update changes.h and tags for release
