+++ doc/users/doxygen.html
@@ -6,7 +6,7 @@
   <title>deal.II Doxygen documentation</title>
   <link href="../screen.css" rel="StyleSheet">
   <meta name="author" content="the deal.II authors">
-  <meta name="copyright" content="Copyright (C) 2012, 2013 by the deal.II authors">
+  <meta name="copyright" content="Copyright (C) 2012 - 2014, 2013 by the deal.II authors">
   <meta name="date" content="$Date$">
   <meta name="keywords" content="deal dealii finite elements fem triangulation">
   <meta http-equiv="content-language" content="en">
@@ -31,6 +31,7 @@
       Doxygen, if the online documentation is available,
       <a href="../doxygen/deal.II/index.html">here</a>. For older revisions,
 	links are provided here by revision number:
+	<a href="http://www.dealii.org/8.2.0/doxygen/deal.tag">8.2.0</a>,
 	<a href="http://www.dealii.org/8.1.0/doxygen/deal.tag">8.1.0</a>,
 	<a href="http://www.dealii.org/8.0.0/doxygen/deal.tag">8.0.0</a>,
 	<a href="http://www.dealii.org/7.3.0/doxygen/deal.tag">7.3.0</a>,
